Fixed some bugs and .h added

+++ src/ft_memcmp.c
@@ -6,13 +6,13 @@
 /*   By: alfgarci <alfgarci@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/07 11:17:34 by alfgarci          #+#    #+#             */
-/*   Updated: 2022/08/07 11:23:26 by alfgarci         ###   ########.fr       */
+/*   Updated: 2022/08/07 15:30:53 by alfgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] == s1[i] && i < n)
