+++ src/ft_strlcat.c
@@ -6,18 +6,18 @@
 /*   By: alfgarci <alfgarci@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/31 22:58:04 by alfgarci          #+#    #+#             */
-/*   Updated: 2022/07/31 23:02:25 by alfgarci         ###   ########.fr       */
+/*   Updated: 2022/08/07 18:33:46 by alfgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "../includes/libft.h"
 
 size_t	ft_strlcat(char *restrict dst, char *restrict src, size_t dstsize)
 {
-	unsigned int	srclen;
-	unsigned int	dstlen;
-	unsigned int	i;
-	unsigned int	j;
+	size_t	srclen;
+	size_t	dstlen;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	srclen = ft_strlen(src);
