+++ src/ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: alfgarci <alfgarci@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/31 22:54:28 by alfgarci          #+#    #+#             */
-/*   Updated: 2022/07/31 22:57:25 by alfgarci         ###   ########.fr       */
+/*   Updated: 2022/08/07 15:41:24 by alfgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 size_t	ft_strlcpy(char *restrict dst, char *restrict src, size_t dstsize)
 {
-	unsigned int	i;
-	unsigned int	len;
+	size_t	i;
+	size_t	len;
 
 	i = 0;
 	len = 0;
