+++ src/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: alfgarci <alfgarci@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/06 13:58:36 by alfgarci          #+#    #+#             */
-/*   Updated: 2022/08/06 14:00:49 by alfgarci         ###   ########.fr       */
+/*   Updated: 2022/08/07 15:44:03 by alfgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while ((s1[i] != '\0' || s2[i] != '\0') && i < n)
