+++ src/ft_strnstr.c
@@ -6,14 +6,14 @@
 /*   By: alfgarci <alfgarci@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/07 11:25:34 by alfgarci          #+#    #+#             */
-/*   Updated: 2022/08/07 11:30:52 by alfgarci         ###   ########.fr       */
+/*   Updated: 2022/08/07 15:47:04 by alfgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
